% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIAParameters.R
\name{add.encode.data}
\alias{add.encode.data}
\title{Retrieve whatever data it can from ENCODE and add it to the ChIA parameter object.}
\usage{
add.encode.data(chia.params)
}
\arguments{
\item{chia.params}{The ChIA parameters object for which additional annotations must be fetched from ENCODE.}
}
\value{
An environment that can be passed to process.chia.pet, annotate.chia.pet or analyze.chia.pet.
}
\description{
Retrieve whatever data it can from ENCODE and add it to the ChIA parameter object.
}

