% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{analyze.annotation}
\alias{analyze.annotation}
\title{Analyze the annotation of ChIA-PET data}
\usage{
analyze.annotation(chia.obj, output.dir = "output")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}}

\item{output.dir}{The name of the directory where to save the graphs.}
}
\description{
Analyzes ChIA-PET data and produces graphs according to the annotation: \describe{
  \item{Proportion of genomic location as a function of connectivity category.pdf} {A plot of the proportion of the genomic location of the regions as a fonction of connectivity category.}
  \item{Contact heatmap for genomic location.pdf} {A contact heatmap of the genomic location of the regions.}}
}

