% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{analyze.expression}
\alias{analyze.expression}
\title{Analyze the expression of ChIA-PET data}
\usage{
analyze.expression(chia.obj, output.dir = "output")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{output.dir}{The name of the directory where to save the graphs.}
}
\description{
Produce a plot to the expression of a node according to its degree.
}

