% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{analyze.gene.specificity}
\alias{analyze.gene.specificity}
\title{Analyze the gene specificity of ChIA-PET data}
\usage{
analyze.gene.specificity(chia.obj, output.dir = "output")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{output.dir}{The name of the directory where to save the graphs.}
}
\description{
Analyzes ChIA-PET data and produces graphs according to gene specificity: \describe{
\item{Tau vs degree.pdf}{A plot of the Tau of the nodes according to their degree.}
\item{Boxplot of degrees by expression category.pdf}{A boxplot the degrees of the nodes in each expression category.}}
}

