% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeChIA.R
\name{analyze.tf}
\alias{analyze.tf}
\title{Analyze the TF of ChIA-PET data
Produce a plot of the presence of TF at the connection points of ChIA-PET data, according to the
connectivity of the nodes.}
\usage{
analyze.tf(chia.obj, output.dir = "output")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{output.dir}{The name of the directory where to save the graphs.}
}
\description{
Analyze the TF of ChIA-PET data
Produce a plot of the presence of TF at the connection points of ChIA-PET data, according to the
connectivity of the nodes.
}

