% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateChIA.R
\name{annotate.chia}
\alias{annotate.chia}
\title{Annotate "chia.obj", given as parameter.}
\usage{
annotate.chia(chia.obj, output.dir, input.chrom.state = NULL,
  tf.regions = NULL, histone.regions = NULL, pol.regions = NULL,
  expression.levels = NULL, genome.build = c("hg19", "mm9", "mm10", "hg38"),
  tssRegion = c(-3000, 3000), centrality.measures = c("Degree"),
  weight.attr = NULL, verbose = TRUE)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{output.dir}{The name of the directory where to write the selected annotations.}

\item{input.chrom.state}{The name of the file containing the information about chromatin states.}

\item{tf.regions}{A \linkS4class{GRangesList} object containing the TF regions.}

\item{histone.regions}{A \linkS4class{GRangesList} object containing the histone regions.}

\item{pol.regions}{A \linkS4class{GRangesList} object containing the pol2 regions.}

\item{expression.levels}{A data frame containing the levels of expression of genes. according to their EMSEMBL id.}

\item{genome.build}{The name of the chosen annotation ("hg38", "mm9", "mm10", "hg19").}

\item{tssRegion}{A vector with the region range to TSS.}

\item{centrality.measures}{The metrics to be sued for assessing node centrality. See \code{\link{associate.centralities}}.}

\item{weight.attr}{The name of the edge attribute to be used as weight, if any.}

\item{verbose}{If TRUE, console output is suppressed.}
}
\value{
The annotated "\code{chia.obj}".
}
\description{
Annotate "chia.obj", given as parameter.
}

