% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{annotate.chip}
\alias{annotate.chip}
\title{Annotate "chip.data", given as parameter.}
\usage{
annotate.chip(chip.data, input.chrom.state, tf.regions, histone.regions,
  expression.levels, genome.build = c("hg19", "mm9", "mm10", "hg38"),
  biosample = "GM12878", tssRegions = c(-3000, 3000), output.dir, label)
}
\arguments{
\item{chip.data}{A \linkS4class{GRanges} object containing the ChIp-seq data.}

\item{input.chrom.state}{The name of the file containing the information about chromatin states.}

\item{tf.regions}{A data frame containing the TF data.}

\item{histone.regions}{A data frame containing the histone data.}

\item{expression.levels}{A data frame containing the levels of expression of genes. according to their EMSEMBL id.}

\item{genome.build}{The name of the chosen annotation ("hg38", "mm9", "mm10", "hg19").}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562.}

\item{output.dir}{The name of the directory where to write the selected annotations.}

\item{label}{The name of the file containing all annotation.}

\item{tssRegion}{A vector with the region range to TSS.}
}
\value{
The annotated "\code{chip.data}".
}
\description{
Annotate "chip.data", given as parameter.
}

