% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{annotate.region}
\alias{annotate.region}
\title{Annotates a set of regions with genomic features.}
\usage{
annotate.region(region, annotations.list, filename = NULL)
}
\arguments{
\item{region}{The regions to annotate.}

\item{annotations.list}{A list of annotation databases returned by 
select.annotations.}

\item{filename}{The name of the file where the results should be saved.}
}
\value{
An annotation object.
}
\description{
Given a genome build identifier, creates a list of annotation databases
which can be passed to the annotation functions.
}

