% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{annotate_region}
\alias{annotate_region}
\title{Annotates a set of regions with genomic features.}
\usage{
annotate_region(region, annotations.list, tssRegion = c(-3000, 3000),
  filename = NULL)
}
\arguments{
\item{region}{The regions to annotate.}

\item{annotations.list}{A list of annotation databases returned by
\code{\link{select_annotations}}.}

\item{tssRegion}{A vector with the region range to TSS.}

\item{filename}{The name of the file where the results should be saved.
If \code{NULL}, results are not saved to disk.}
}
\value{
An annotation object.
}
\description{
Given a genome build identifier, creates a list of annotation databases
which can be passed to the annotation functions.
}

