% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{annotate_venn_center}
\alias{annotate_venn_center}
\title{Annotate the inner group of an \code{intersect.object}.}
\usage{
annotate_venn_center(intersect.object, annotations.list, filename = NULL)
}
\arguments{
\item{intersect.object}{An intersect object returned by \code{\link{build_intersect}}.}

\item{annotations.list}{A list of annotation objects returned by \code{\link{select_annotations}}}

\item{filename}{A filename for the resulting annotations. Pass \code{NULL}
to skip saving to a file.}
}
\value{
The annotations for the intersect's inner regions.
}
\description{
Given an \code{intersect.object}, generate annotations for the regions
where all factors are present.
}
