% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{annotate_venn_exclusive}
\alias{annotate_venn_exclusive}
\title{Annotate the outer groups of an \code{intersect.object}.}
\usage{
annotate_venn_exclusive(intersect.object, annotations.list,
  file.prefix = NULL)
}
\arguments{
\item{intersect.object}{An intersect object returned by \code{\link{build_intersect}}.}

\item{annotations.list}{A list of annotation objects returned by \code{\link{select_annotations}}}

\item{file.prefix}{A prefix for the names of the output files where the
annotations will be written. Pass \code{NULL} to skip saving to a file.}
}
\value{
A list with the generated annotations.
}
\description{
Given an \code{intersect.object}, generate annotations for the regions
where only one factor is present. One annotation per factor is generated.
}
