% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate.chrom.state}
\alias{associate.chrom.state}
\title{Associate chromatin sates to a \linkS4class{GRanges} object.}
\usage{
associate.chrom.state(regions, input.chrom.state)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{input.chrom.state}{The path to a bed file representing a genome
segmentation by chromatin state.}
}
\value{
The \linkS4class{GRanges} object with associated chromatin states.
}
\description{
Associate chromatin sates to a \linkS4class{GRanges} object.
}

