% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateChIA.R
\name{associate.components}
\alias{associate.components}
\title{Associates components ids and sizes to chia data, as returned by \code{\link{load.chia}}.}
\usage{
associate.components(chia.obj)
}
\arguments{
\item{chia.obj}{ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{split}{Should the data be divided into communities?}

\item{oneByOne}{Sould the netwoks be treated one by one or as a whole?}

\item{method}{What method sould be used to split data (ignored if split = \code{FALSE}).}
}
\value{
The annotated chia.obj.
}
\description{
Associates components ids and sizes to chia data, as returned by \code{\link{load.chia}}.
}

