% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateChIA.R
\name{associate.gene}
\alias{associate.gene}
\title{Associate genes to a \linkS4class{GRanges} object from ChIA-PET data.}
\usage{
associate.gene(regions, expression.data = NULL)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{expression.data}{A data frame containing the levels of expression of genes, according to their EMSEMBL id.}
}
\value{
The \linkS4class{GRanges} object with associated genes.
}
\description{
Associate genes to a \linkS4class{GRanges} object from ChIA-PET data.
}

