% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{associate.genomic.region}
\alias{associate.genomic.region}
\title{Associate genomic regions with the \code{Regions} of "chia.obj".}
\usage{
associate.genomic.region(chia.obj, genome.build, output.dir)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by load.chia.}

\item{genome.build}{The name of the chosen annotation ("hg38", "mm9", "mm10", "hg19").}

\item{output.dir}{The name of the directory where to write the selected annotations.}
}
\value{
"\code{chia.obj}" with associated genomic regions.
}
\description{
Associate genomic regions with the \code{Regions} of "chia.obj".
}

