% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate.is.gene.active}
\alias{associate.is.gene.active}
\title{Associates a boolean in fonction of the activity of the gene}
\usage{
associate.is.gene.active(regions)
}
\description{
Is.Gene.Active is \code{TRUE} if FPKM > 1
}

