% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateChIA.R
\name{associate.is.in.factory}
\alias{associate.is.in.factory}
\title{Associates boolean to regions in fonction of their presence in factories}
\usage{
associate.is.in.factory(regions)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}
}
\value{
The annotated regions.
}
\description{
Is.In.Factory is \code{TRUE} if the region is in a network with 3 genes or more.
}

