% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{associate.tf}
\alias{associate.tf}
\title{Associate TF overlaps with the \code{Regions} of "chia.obj".}
\usage{
associate.tf(chia.obj, tf.regions)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{tf.regions}{A data frame containing the TF data.}
}
\value{
"\code{chia.obj}" with associated TF overlaps.
}
\description{
Associate TF overlaps with the \code{Regions} of "chia.obj".
}

