% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{associate.tissue.specificity.human}
\alias{associate.tissue.specificity.human}
\title{Associate tissue specificity of genes with the \code{Regions} of "chia.obj".}
\usage{
associate.tissue.specificity.human(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
"\code{chia.obj}" with associated tissue scpecificity.
}
\description{
Associate tissue specificity of genes with the \code{Regions} of "chia.obj".
}

