% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate_gene_chip}
\alias{associate_gene_chip}
\title{Associate genes to a \linkS4class{GRanges} object from ChIP-seq data.}
\usage{
associate_gene_chip(regions, expression.data = NULL, biosample, genome.build)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{expression.data}{A data frame containing the levels of expression of genes, according to their EMSEMBL id.}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or MCF-7.}

\item{genome.build}{Which genome assembly should the results come from?}
}
\value{
The \linkS4class{GRanges} object with associated genes.
}
\description{
Associate genes to a \linkS4class{GRanges} object from ChIP-seq data.
}

