% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate_genomic_region}
\alias{associate_genomic_region}
\title{Associate genomic regions to a \linkS4class{GRanges} object.}
\usage{
associate_genomic_region(regions, genome.build, output.dir,
  tssRegion = c(-3000, 3000))
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{genome.build}{The name of the chosen annotation ("hg38", "mm9", "mm10", "hg19").}

\item{output.dir}{The name of the directory where to write the selected annotations.}

\item{tssRegion}{A vector with the region range to TSS.}
}
\value{
The \linkS4class{GRanges} object with associated genomic regions.
}
\description{
Associate genomic regions to a \linkS4class{GRanges} object.
}

