% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate_is_gene_active}
\alias{associate_is_gene_active}
\title{Associates a boolean in fonction of the activity of the gene}
\usage{
associate_is_gene_active(regions)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}
}
\value{
The \linkS4class{GRanges} object with associated histone overlap percentage.
}
\description{
Is.Gene.Active is \code{TRUE} if FPKM > 1
Associate gene activity marker to a \linkS4class{GRanges} object.
}

