% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate_tf}
\alias{associate_tf}
\title{Associate TF overlaps to a \linkS4class{GRanges} object.}
\usage{
associate_tf(regions, tf.regions)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{tf.regions}{A data frame containing the TF data.}
}
\value{
The \linkS4class{GRanges} object with associated TF overlaps.
}
\description{
Associate TF overlaps to a \linkS4class{GRanges} object.
}
