% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate_tissue_specificity_human}
\alias{associate_tissue_specificity_human}
\title{Associate tissue specificity of genes to a \linkS4class{GRanges} object.}
\usage{
associate_tissue_specificity_human(regions)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}
}
\value{
The \linkS4class{GRanges} object with associated tissue scpecificity.
}
\description{
Associate tissue specificity of genes to a \linkS4class{GRanges} object.
}
