% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{average.component.size}
\alias{average.component.size}
\title{Return the mean component size of a CHIA object.}
\usage{
average.component.size(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
The mean component size of the chia object.
}
\description{
Return the mean component size of a CHIA object.
}

