% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeChIA.R
\name{boxplot.by.connectivity}
\alias{boxplot.by.connectivity}
\title{Boxplot by connectivity}
\usage{
boxplot.by.connectivity(chia.obj, variable.name, label, output.dir)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{variable.name}{The name of the variable according to which the boxplot should be computed.}

\item{label}{The Name to give te the variable name in the resulting heatmap.}

\item{output.dir}{The name of the directory where to save the heatmaps.}
}
\description{
Creates a boxplot of a specific value (given as parameter) according to the connectivity
}

