% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{boxplot.per.tf}
\alias{boxplot.per.tf}
\title{Produces boxplots for every transcription factor, in relation to its 3D connectivity}
\usage{
boxplot.per.tf(chip.data, biosample, genome.build, chia.obj, output.dir,
  TSS = TRUE, tssRegion = c(-3000, 3000))
}
\arguments{
\item{chip.data}{A \linkS4class{GRangesList} containing the regions of the ChIp-seq data, with signal values.}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are GM12878, K562 or MCF-7.}

\item{genome.build}{The name of the chosen annotation ("hg38", "hg19").}

\item{chia.obj}{Annotated ChIA-PET data, as returned by \link{analyze.chia.pet} or \link{annotate.chia}.}

\item{output.dir}{The directory where to write the boxplots.}

\item{TSS}{Should only the TSS regions be kept?}

\item{tssRegion}{tssRegion A vector with the region range to TSS.}
}
\description{
For every transcription factor in the ChIP data, creates a boxplot of the force of the ChIP-seq signal
in function of the contact frequence of the region.
}

