% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIAParameters.R
\name{build.chia.params}
\alias{build.chia.params}
\title{Create an object storing parameters for ChIA processing.}
\usage{
build.chia.params(input.chrom.state = NULL, biosample = NULL,
  genome.build = NULL, tf.regions = NULL, histone.regions = NULL,
  pol.regions = NULL, expression.data = NULL, tssRegion = c(-3000, 3000),
  centrality.measures = c("Degree"), weight.attr = NULL)
}
\arguments{
\item{input.chrom.state}{The name of the file containing the information about chromatin states.}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are GM12878, K562 or MCF-7.}

\item{genome.build}{The name of the chosen annotation ("hg38", "hg19").}

\item{tf.regions}{A \linkS4class{GRangesList} object containing TF regions.}

\item{histone.regions}{A \linkS4class{GRangesList} object containing histone regions.}

\item{pol.regions}{A \linkS4class{GRangesList} object containing the pol2 regions.}

\item{expression.levels}{A data frame containing the levels of expression of genes, according to their EMSEMBL id.}
}
\value{
An environment that can be passed to process.chia.pet, annotate.chia.pet or analyze.chia.pet.
}
\description{
Create an object storing parameters for ChIA processing.
}

