% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{build.intersect.all}
\alias{build.intersect.all}
\title{Builds an overlap of the given regions and output all default annotations.}
\usage{
build.intersect.all(regions, annotations.list, label)
}
\arguments{
\item{regions}{A GRangesList of the regions to be intersected and analyzed.}

\item{annotations.list}{A list of annotation objects returned by select.annotations}

\item{label}{A label to use when generating file names.}
}
\value{
The generated intersect object.
}
\description{
After building the overlap, this function will generate:
 - A venn diagram.
 - An annotation of the intersect's inner regions.
 - An annotation of the intersect's outer regions.
 - The pairwise overlap of all factors.
}

