% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{build_intersect_all}
\alias{build_intersect_all}
\title{Builds an overlap of the given regions and output all default annotations.}
\usage{
build_intersect_all(regions, annotations.list, label)
}
\arguments{
\item{regions}{A \linkS4class{GRangesList} of the regions to be intersected and analyzed.}

\item{annotations.list}{A list of annotation objects returned by \code{\link{select_annotations}}}

\item{label}{A label to use when generating file names.}
}
\value{
The generated intersect object.
}
\description{
After building the overlap, this function will generate: \enumerate{
  \item A venn diagram.
  \item An annotation of the intersect's inner regions.
  \item An annotation of the intersect's outer regions.
  \item The pairwise overlap of all factors.}
}
