% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateChIA.R
\name{calculate.centralities}
\alias{calculate.centralities}
\title{Calculates centralities for all vertices in a graph.}
\usage{
calculate.centralities(chia.obj, which.measures = c("Degree", "Betweenness",
  "Eigenvector"), weight.attr = NULL)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{which.measures}{A vector containing the names of the measures to be used
to assess centrality. Those can be "Degree", "Betweenness" and "Eigenvector".}

\item{weight.attr}{The anme of the edge attribute to be sued as a weight in 
centrality calculations.}
}
\value{
A data-frame containing the centrality scores and markers for all vertices in the graph.
}
\description{
Calculates centralities for all vertices in a graph.
}

