% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{characterize_gene_set}
\alias{characterize_gene_set}
\title{Given a set of Entrez gene ids, perform  KEGG enrichment and motif enrichement at the promoters.}
\usage{
characterize_gene_set(gene.set, annotations.list, skip.motif = FALSE,
  skip.kegg = FALSE, output.dir = "output/", promoter.size = 1000)
}
\arguments{
\item{gene.set}{A vector of Entrez IDs representing the genes to be characterized.}

\item{annotations.list}{A list of annotation databases returned by
\code{\link{select_annotations}}.}

\item{skip.motif}{If \code{TRUE}, motif enrichment is skipped.}

\item{skip.kegg}{If \code{TRUE}, KEGG pathway enrichment is skipped.}

\item{output.dir}{The directory where output should be stored.A directory for storing output.}

\item{file.label}{A label for generating file names to save the results.
If \code{NULL}, results are not saved to disk.}
}
\value{
A list containing the characterization results.
}
\description{
Given a set of Entrez gene ids, perform  KEGG enrichment and motif enrichement at the promoters.
}
