% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{chia.left}
\alias{chia.left}
\title{Return the left part of the ChIA-PET data.}
\usage{
chia.left(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A \linkS4class{GRanges} object with the \code{Regions} from the "\code{chia.obj}" parameter corresponding the the "left side".
of the original data.
}
\description{
Return the left part of the ChIA-PET data.
}

