% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{chrom_state_match}
\alias{chrom_state_match}
\title{Given a set of regions and a chromatin state map, match the "best" chromatin state to each region.}
\usage{
chrom_state_match(regions, chrom.states.gr)
}
\arguments{
\item{regions}{The regions to which chromatin states should be associated.}

\item{chrom.states.gr}{A GRanges containing the chromatin state map.}
}
\value{
A vector of the same length as regions, indicating the "best" chromatin state mapping to that region.
}
\description{
Given a set of regions and a chromatin state map, match the "best" chromatin state to each region.
}

