% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{collapse.regions}
\alias{collapse.regions}
\title{Collapses a list of genomic ranges into a single set of unique, 
non-overlapping ranges.}
\usage{
collapse.regions(gr.list)
}
\arguments{
\item{gr.list}{The ranges to be collapsed.}
}
\value{
The collapsed regions.
}
\description{
Ranges are prioritized in the input list order. So, if the first element
of the list (A) covers the range 1-10, and the second element (B) covers 
the range 5-15, then the resulting ranges will have a 1-10 range named A,
and a 11-15 range named 'B'.
}

