% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeChIA.R
\name{connectivity.enrichment}
\alias{connectivity.enrichment}
\title{Perform enrichments by connectivity group.}
\usage{
connectivity.enrichment(chia.obj, variable.name, label, wrap.row = 3,
  wrap.col = 6, output.dir = "output")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{variable.name}{The name of the variable according to which the enrichment should be performed.}

\item{label}{The Name to give te the variable name in the resulting graph.}

\item{wrap.row}{The number of rows in the final file.}

\item{wrap.col}{The number of columns in the final file.}

\item{output.dir}{The name of the directory where to save the graphs.}
}
\value{
The plots of enrichments by connectivity group.
}
\description{
Perform enrichments by connectivity group.
}

