% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{contact.heatmap}
\alias{contact.heatmap}
\title{Create a heatmap by connectivity group.}
\usage{
contact.heatmap(chia.obj, variable.name, label, output.dir)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{variable.name}{The name of the variable according to which the heatmap should be computed.}

\item{label}{The Name to give te the variable name in the resulting heatmap.}

\item{output.dir}{The name of the directory where to save the heatmaps.}
}
\value{
The matrix used to create the heatmap.
}
\description{
Create a heatmap by connectivity group.
}

