% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsawAnalysis.R
\name{csaw_analyze}
\alias{csaw_analyze}
\title{Analyze bam files with csaw and produces corresponding graphs.}
\usage{
csaw_analyze(correspondances, reference, genome.build,
  output.dir = "csaw_output/", threshold = 3, save.plots = TRUE)
}
\arguments{
\item{correspondances}{A data frame with two columns: \describe{
\item{$Name}{Name (with the path) of the bam files.}
\item{$Factors}{The factors to considerate during the analysis. Files with the same factor are replicas.}}}

\item{reference}{The name of the reference factor.}

\item{genome.build}{The name of the chosen annotation ("hg38", "hg19", "mm9", "mm10").}

\item{output.dir}{The name of the directory where to save the files and graphs.}

\item{threshold}{The threshold to consider while filtering data.}

\item{save.plots}{Should plots and files created by \code{\link{csaw.save.plots}} be saved?}
}
\description{
Analyze bam files with csaw and produces corresponding graphs.
}

