% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{download.chromatin.states}
\alias{download.chromatin.states}
\title{Helper function for obtaining chromatin states through import.chrom.states.}
\usage{
download.chromatin.states(biosample, assembly,
  download.dir = file.path("input/ENCODE", biosample, assembly))
}
\arguments{
\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or MCF-7.}

\item{download.dir}{The folder where the downloaded files should be stored.
defaults to \code{file.path("input/ENCODE", biosample, assembly, "chip-seq", "tf")}.}

\item{genome.assembly}{Which genome assembly should the results come from?}
}
\value{
A GRanges object with the loaded chromatin states, or NULL if the do not exist.
}
\description{
Helper function for obtaining chromatin states through import.chrom.states.
}

