% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{download.encode.chip}
\alias{download.encode.chip}
\title{Obtains and process transcription factor data from ENCODE.}
\usage{
download.encode.chip(biosample, assembly,
  download.filter = default.download.filter.chip,
  download.dir = file.path("input/ENCODE", biosample, "chip-seq"),
  signal.value = FALSE)
}
\arguments{
\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or MCF-7.}

\item{download.filter}{A filtering function to be applied to the experiment
data-frame returned by \code{\link[ENCODExplorer]{queryEncode}}. The function should return a filtered
\code{data-frame} containing only the files which should be downloaded.}

\item{download.dir}{The folder where the downloaded files should be stored.
defaults to \code{file.path("input/ENCODE", biosample, "chip-seq")}.}

\item{signal.value}{Should the signalValue from the ChIP-seq be kept?}

\item{genome.assembly}{Which genome assembly should the results come from?}
}
\value{
A list containing three elements: \describe{
  \item{Metadata} {The metadata returned by \code{\link[ENCODExplorer]{queryEncode}}, containing information
    about all files which matched the query.}
  \item{Downloaded} {The list of files which were downloaded.}
  \item{Regions} {The processed peak regions.}}
}
\description{
Given an encode biosample identifier (GM12878, MCF-7, etc.), this function
retrieved processed bed files giving the genomic location of binding peaks.
}

