% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{download.encode.rna}
\alias{download.encode.rna}
\title{Obtains and process expression data from ENCODE.}
\usage{
download.encode.rna(biosample, assembly,
  download.filter = default.download.filter.rna,
  download.dir = file.path("input/ENCODE", biosample, "rna-seq"))
}
\arguments{
\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or MCF-7.}

\item{download.filter}{A filtering function to be applied to the experiment
data-frame returned by \code{\link[ENCODExplorer]{queryEncode}}. The function should return a filtered
\code{data-frame} containing only the files which should be downloaded.}

\item{download.dir}{The folder where the downloaded files should be stored.
defaults to \code{file.path("input/ENCODE", biosample, "chip-seq")}.}

\item{genome.assembly}{Which genome assembly should the results come from?}
}
\value{
A list containing three elements: \describe{
  \item{Metadata} {The metadata returned by queryEncode, containing information
    about all files which matched the query.}
  \item{Downloaded} {The list of files which were downloaded.}
  \item{Expression} {A data-frame with processed expression levels.}}
}
\description{
Given an encode biosample identifier (GM12878, MCF-7, etc.), this function
retrieved processed expression levels.
}

