% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{download.one.chrom.state}
\alias{download.one.chrom.state}
\title{Download a single chromatin state map from ENCODE.}
\usage{
download.one.chrom.state(mnemonic, number.of.states, download.dir)
}
\arguments{
\item{mnemonic}{The ENCODE mnemonic for the cell type (for example, "E123")}

\item{number.of.states}{Which chromatin state map to download, either 15 or 18.}

\item{download.dir}{The folder where the downloaded files should be stored.}
}
\value{
The path of the downloaded file, if it was found.
}
\description{
Download a single chromatin state map from ENCODE.
}

