% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{download_encode_polymerases}
\alias{download_encode_polymerases}
\title{Helper function for obtaining polymerase data through download_encode_chip.}
\usage{
download_encode_polymerases(biosample, assembly,
  download.dir = file.path("input/ENCODE", biosample, assembly, "chip-seq",
  "polymerases"), ...)
}
\arguments{
\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or MCF-7.}

\item{download.dir}{The folder where the downloaded files should be stored.
defaults to \code{file.path("input/ENCODE", biosample, assembly, "chip-seq", "tf")}.}

\item{...}{Other parameters to be passed to download_encode_chip.}

\item{genome.assembly}{Which genome assembly should the results come from?}
}
\value{
A list containing three elements: \describe{
\item{Metadata}{The metadata returned by \code{\link[ENCODExplorer]{queryEncode}}, containing information
    about all files which matched the query.}
\item{Downloaded}{The list of files which were downloaded.}
\item{Regions}{The processed peak regions.}}
}
\description{
Helper function for obtaining polymerase data through download_encode_chip.
}
