% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsawAnalysis.R
\name{filter.reads.to.counts}
\alias{filter.reads.to.counts}
\title{Filter out unintersing windows, by global enrichment method.}
\usage{
filter.reads.to.counts(bam.files, param, counted.reads, threshold)
}
\arguments{
\item{bam.files}{A vector with the path to bam files.}

\item{param}{A vector with the list of parameters to give to the functions.}

\item{counted.reads}{A \linkS4class{RangedSummarizedExperiment} object returned by \code{\link[csaw]{WindowCounts}}.}

\item{threshold}{A numeric, the threshold to consider in order to cut out windows.}
}
\value{
A list with the following elements: \describe{
\item{[[1]]}{A filtered \linkS4class{RangedSummarizedExperiment} object returned by \code{\link[csaw]{WindowCounts}}.}
\item{[[2]]}{filter.stat, list, statistics of windows aboundances and their filter}}
}
\description{
Filter out unintersing windows, by global enrichment method.
}

