% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{gene.from.regions}
\alias{gene.from.regions}
\title{Given a set of regions, convert it to a set of Entrez gene ids.}
\usage{
gene.from.regions(regions, annotations.list, flank.size = c(-3000, 3000),
  region.types = c("Gene body", "Promoter", "Downstream", "Distal", "All"))
}
\arguments{
\item{regions}{A GRanges object with regions to be associated to genes.}

\item{annotations.list}{A list of annotation databases returned by 
select.annotations.}

\item{flank.size}{The extent around the TSS which is considered the 
promoter region.}

\item{region.types}{A character vector representing region types which will
cause a gene association. These can be:
  - *Promoter*: Promoter region (upstream and downstream of the TSS)
  - *Gene body*: Exons, introns, 5' and 3' UTRs.
  - *Downstream*: Region downstream of the TES.
  - *Distal*: Regions which do not fit any of the above category.
  - *All*: All of the above.
disease pathways.}
}
\value{
A data-frame with the enrichment results.
}
\description{
Utility function to get a gene list from a set of regions.
}

