% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{gene.peaks.distances}
\alias{gene.peaks.distances}
\title{Generate a summary of the distances separating a set of genes from a set of peaks.}
\usage{
gene.peaks.distances(peaks, genes, annot)
}
\arguments{
\item{peaks}{A GRanges object describing the peaks whose distance from the genes must be summarized.}

\item{genes}{A vector of gene symbols.}

\item{annot}{An annotation list returned by select.annotation for putting th esymbols in context.}
}
\value{
A numeric vector containing the proportion of genes within each distance category.
}
\description{
Generate a summary of the distances separating a set of genes from a set of peaks.
}

