% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{gene_from_regions}
\alias{gene_from_regions}
\title{Given a set of regions, convert it to a set of Entrez gene ids.}
\usage{
gene_from_regions(regions, annotations.list, flank.size = c(-3000, 3000),
  region.types = c("Gene body", "Promoter", "Downstream", "Distal", "All"))
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object with regions to be associated to genes.}

\item{annotations.list}{A list of annotation databases returned by
\code{\link{select_annotations}}.}

\item{flank.size}{The extent around the TSS which is considered the
promoter region.}

\item{region.types}{A character vector representing region types which will
  cause a gene association. These can be: \describe{
\item{Promoter}{Promoter region (upstream and downstream of the TSS).}
\item{Gene body}{Promoter, Exons, introns, 5' and 3' UTRs.}
\item{Downstream}{Region downstream of the TES.}
\item{Distal}{Regions which do not fit any of the above category.}
\item{All}{All of the above.}}}
}
\value{
A data-frame with the enrichment results.
}
\description{
Utility function to get a gene list from a set of regions.
}

