% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{genes.by.component}
\alias{genes.by.component}
\title{Return the number genes per component in the CHIA object.}
\usage{
genes.by.component(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
The number of active genes represented in the chia object.
}
\description{
Return the number genes per component in the CHIA object.
}

