% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{get_encode_mnemonic}
\alias{get_encode_mnemonic}
\title{Given a biosample's name, tries to match it with an ENCODE mnemonic.}
\usage{
get_encode_mnemonic(biosample)
}
\arguments{
\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or HeLa.}
}
\value{
The ENCODE mnemonic associated with the biosample.
}
\description{
Given a biosample's name, tries to match it with an ENCODE mnemonic.
}

