% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{get_promoters}
\alias{get_promoters}
\title{Given a set of Entrez IDS, retrieve the coordinates of those genes' promoters.}
\usage{
get_promoters(selected.genes, annotations.list, flank.size = 1000)
}
\arguments{
\item{selected.genes}{A vector of ENTREZ gene ids whose promoters should be
retrieved.}

\item{annotations.list}{A list of annotation databases returned by
\code{\link{select_annotations}}.}

\item{flank.size}{How many base pairs upstream of the TSS should we retrieve?}
}
\value{
A \linkS4class{GRanges} objects representing the promoters of the given genes.
}
\description{
Given a set of Entrez IDS, retrieve the coordinates of those genes' promoters.
}

