% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{histone.download.filter.chip}
\alias{histone.download.filter.chip}
\title{Alternative filtering function for \code{\link{download.encode.chip}}.}
\usage{
histone.download.filter.chip(query.results, genome.assembly)
}
\arguments{
\item{query.results}{A partial \code{data.frame} obtained from the \code{\link[ENCODExplorer]{queryEncode}}
function.The biosample identifier from ENCODE. Valid examples are GM12878, K562.}

\item{genome.assembly}{Which genome assembly should the results come from?}
}
\value{
A filtered \code{data frame}.
}
\description{
The filtering function does three things: \enumerate{
  \item It removes all files which do not have the correct genome assembly.
  \item It removes all marks chips, except histone marks chips.
  \item If the provided results have been re-analyzed by the ENCODE Consortium,
    the ENCODE results are kept and the original ones discarded.}
}

