% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{identify.crossing.edges}
\alias{identify.crossing.edges}
\title{Identifies edges crossing community borders.}
\usage{
\method{identify}{crossing.edges}(input.graph,
  method = igraph::cluster_fast_greedy, weight.attr = NULL)
}
\arguments{
\item{input.graph}{The igraph whose community-crossing edges must be identified,}

\item{method}{A method returning a \code{community} object, such as igraph::cluster_fast_greedy.}

\item{weight.attr}{The name fo the edge attribute to be used as edge weight.}
}
\value{
The ids of the edges to be removed.
}
\description{
Identifies edges crossing community borders.
}

