% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportHelper.R
\name{import.and.discard.metadata}
\alias{import.and.discard.metadata}
\title{Import a bed file and discards its metadata.}
\usage{
import.and.discard.metadata(x, extraCols)
}
\arguments{
\item{x}{The name of the BED file to import.}

\item{extraCols}{The names and types of the extra columns.}
}
\value{
A \linkS4class{GRanges} object representing the regions of x, without the metadata.
}
\description{
Import a bed file and discards its metadata.
}
