% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportHelper.R
\name{import.files.into.grl}
\alias{import.files.into.grl}
\title{Import a set of bed-lie files into a \linkS4class{GRangesList}.}
\usage{
import.files.into.grl(all.files, file.format, file.ext,
  discard.metadata = FALSE)
}
\arguments{
\item{all.files}{A named vector with the files to be imported. The GRangesList elements
will share the vector element's names.}

\item{file.format}{The format of the files to be imported. Can be "bed",
"broad" or "narrow".}
}
\value{
A \linkS4class{GRanges} object representing the regions of the files in input.dir.
}
\description{
Import a set of bed-lie files into a \linkS4class{GRangesList}.
}

