% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{import_chrom_states}
\alias{import_chrom_states}
\title{Import HMM chromatin states for the right cell type.}
\usage{
import_chrom_states(biosample, download.dir = ".")
}
\arguments{
\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or HeLa.}

\item{download.dir}{The folder where the downloaded files should be stored.}
}
\value{
The name and path of the downloaded and unzipped files.
}
\description{
Import HMM chromatin states for the right cell type.
}

