% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportHelper.R
\name{import_into_grl}
\alias{import_into_grl}
\title{Import a set of bed files in a directory into a \linkS4class{GRangesList}.}
\usage{
import_into_grl(input.dir = ".", file.format = "bed", file.ext = NULL,
  discard.metadata = FALSE, dir.type = "plain")
}
\arguments{
\item{input.dir}{The name of the directory from which BED files must be imported.}

\item{file.format}{The format of the files to be imported. Can be "bed",
"broad" or "narrow".}

\item{file.ext}{The extension of the files to be imported.}

\item{discard.metadata}{If \code{TRUE}, metadata will be discarded after importing.}

\item{dir.type}{The architecture of the directory from which the files are to
be imported. If "plain", files are sought directly within the directory. If
"mugqic", the output structure of the MUGQIC pipeline is searched for.}
}
\value{
A \linkS4class{GRanges} object representing the regions of the files in input.dir.
}
\description{
Import a set of bed files in a directory into a \linkS4class{GRangesList}.
}
