% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{inclusive.overlap.internal}
\alias{inclusive.overlap.internal}
\title{Internal inclusive overlap.}
\usage{
inclusive.overlap.internal(all.regions, overlap.matrix, which.factors)
}
\arguments{
\item{all.regions}{A \linkS4class{GRanges} object with the regions represented}

\item{overlap.matrix}{A matrix, with rows corresponding to regions
in \code{all.rgions}, and columns corresponding to proteins/factors. A
non-zero value in the matrix indicates the protein/factor of interest
is present at this region.}

\item{which.factors}{Indices of columns where the factor should be present.}
}
\value{
A \linkS4class{GRanges} object with the regions matching the given criteria.
}
\description{
Returns the set of regions within \code{all.regions} where all
conditions/proteins at indices which.factors are present,
regardless of whether or not other factors are. Presence/absence
at a given locus is obtained from the overlap.matrix parameter.
}

