% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{install.annotations}
\alias{install.annotations}
\title{Install required annotation packages.}
\usage{
install.annotations(which.assembly)
}
\arguments{
\item{which.assembly}{Which assembly's annotations should be installed.
Supported assemblies are hg19, hg28, mm10 and mm9.}
}
\description{
Install \pkg{BiocInstaller} and the required annotation packages for the given
genome assemblies.
}

