% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{internal.download.chrom.states}
\alias{internal.download.chrom.states}
\title{Attempts to download 15 and 18 chromatin state maps from ENCODE for a given cell type.}
\usage{
internal.download.chrom.states(mnemonic, download.dir = ".")
}
\arguments{
\item{mnemonic}{The ENCODE mnemonic for the cell type (for example, "E123")}

\item{download.dir}{The folder where the downloaded files should be stored.}
}
\value{
A list containing the names of the unzipped chromatin state files, if they were found.
}
\description{
Attempts to download 15 and 18 chromatin state maps from ENCODE for a given cell type.
}

