% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{intersect.overlap}
\alias{intersect.overlap}
\title{Calculate an overlap of certain factors within an intersect.object.}
\usage{
intersect.overlap(intersect.object, indices = NULL, names = NULL,
  exclusive = FALSE)
}
\arguments{
\item{intersect.object}{An intersect.object returned by build.intersect.
by the rows of overlap.matrix.}

\item{indices}{A vector of indices into the factors of intersect.object.}

\item{names}{A vector of factor names in intersect.object.}

\item{exclusive}{If TRUE, a region will be returned if the factors in}
}
\value{
A GRanges object with the regions matching the given criteria.
}
\description{
Given an intersect.object , finds all regions where all
factors described by either indices or names are present. 
If both indices and names are NULL, the inner intersection
is calculated.
}

