% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{intersect_overlap}
\alias{intersect_overlap}
\title{Calculate an overlap of certain factors within an intersect.object.}
\usage{
intersect_overlap(intersect.object, indices = NULL, names = NULL,
  exclusive = FALSE)
}
\arguments{
\item{intersect.object}{An \code{intersect.object} returned by \code{\link{build_intersect}}.
by the rows of \code{overlap.matrix}.}

\item{indices}{A vector of indices into the factors of \code{intersect.object}.}

\item{names}{A vector of factor names in \code{intersect.object}.}

\item{exclusive}{If \code{TRUE}, a region will be returned if the factors in}
}
\value{
A \linkS4class{GRanges} object with the regions matching the given criteria.
}
\description{
Given an \code{intersect.object} , finds all regions where all
factors described by either indices or names are present.
If both indices and names are \code{NULL}, the inner intersection
is calculated.
}

