% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{intersect_venn_plot}
\alias{intersect_venn_plot}
\title{Generates a venn diagram from an \code{intersect.object}.}
\usage{
intersect_venn_plot(intersect.object, filename = NULL)
}
\arguments{
\item{intersect.object}{An intersect object returned by \code{\link{build_intersect}}.}

\item{filename}{A filename for the resulting venn diagram. Pass \code{NULL}
to skip saving to a file.}
}
\value{
The grid object representing the venn.diagram.
}
\description{
Generates a venn diagram from an \code{intersect.object}.
}

