% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{isoform.download.filter.rna}
\alias{isoform.download.filter.rna}
\title{Alternative filtering function for \code{\link{download.encode.rna}}.}
\usage{
isoform.download.filter.rna(query.results, genome.assembly)
}
\arguments{
\item{query.results}{A partial \code{data.frame} obtained from the \code{\link[ENCODExplorer]{queryEncode}}
function.The biosample identifier from ENCODE. Valid examples are GM12878, K562.}

\item{genome.assembly}{Which genome assembly should the results come from?}
}
\value{
A filtered \code{data frame}.
}
\description{
The filtering function does three things: \enumerate{
  \item It removes all files which do not have the correct genome assembly.
  \item It only keeps files with counts for isoforms, not transcripts or genes.
  \item If removes any file where treatment is not NA.}
}

