% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{kegg.enrichment.regions}
\alias{kegg.enrichment.regions}
\title{Perform KEGG enrichment on a set of regions.
 
Conveniance function. Regions are converted to genes with 
gene.from.regions using default parameters.}
\usage{
kegg.enrichment.regions(regions, annotations.list, ...)
}
\arguments{
\item{regions}{A GRanges object with regions to enriched for KEGG pathways.}

\item{annotations.list}{A list of annotation databases returned by 
select.annotations.}

\item{...}{Parameters to be passed to kegg.enrichment.}
}
\value{
A vector of Entrez gene ids containing a non-redundant list of the 
  genes represented by the given regions.
}
\description{
Perform KEGG enrichment on a set of regions.
 
Conveniance function. Regions are converted to genes with 
gene.from.regions using default parameters.
}

