% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{load.chia}
\alias{load.chia}
\title{Read and load a ChIA-PET output file.}
\usage{
load.chia(input.chia)
}
\arguments{
\item{input.chia}{The path of the file containing the ChIA-PET data.}
}
\value{
A list of 4 elements: \describe{
\item{$Left}{A \linkS4class{GRanges} object containing the information about the "left side" of the ChIA-PET data.}
\item{$Right}{A \linkS4class{GRanges} object containing the information about the "right side" of the ChIA-PET data.}
\item{$Regions}{A \linkS4class{GRanges} object containing the reduced information of both sides.}
\item{$Graph}{A directed \linkS4class{igraph} object picturing every interaction between the left side and the right
side of the ChIA-PET data.}}
}
\description{
Read and load a ChIA-PET output file.
}

