% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{load_chrom_state}
\alias{load_chrom_state}
\title{Given the name of a bed file containing a chromatin state map, loads it into a GRanges object.}
\usage{
load_chrom_state(input.chrom.state)
}
\arguments{
\item{input.chrom.state}{The name of the chromatin state bed file.}
}
\value{
A Granges object with the chromatin states.
}
\description{
Given the name of a bed file containing a chromatin state map, loads it into a GRanges object.
}

