% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{motif.enrichment.genes}
\alias{motif.enrichment.genes}
\title{Perform motif enrichment on the promoters of a set of genes.}
\usage{
motif.enrichment.genes(selected.genes, annotations.list, flank.size = 1000,
  ...)
}
\arguments{
\item{selected.genes}{A vector of ENTREZ gene ids whose promoters should be
subjected to motif enrichment.}

\item{annotations.list}{A list of annotation databases returned by 
select.annotations.}

\item{flank.size}{How many base pairs upstream of the TSS should we retrieve?}

\item{...}{Additional arguments for motif.enrichment.}
}
\value{
A list with the following elements:
  - *Region*: The subset of regions where motifs were sought.
  - *Enrichment*: The result of the PWMEnrich motifEnrichment call.
  - *Report*: The result of the PWMEnrich groupReport call.
}
\description{
Utility function which performs the same operations as motif.enrichment,
but accepts a list of genes instead of a list of regions.
}

