% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{motif_enrichment}
\alias{motif_enrichment}
\title{Given a \linkS4class{GRanges} object, finds enriched motifs within those regions.}
\usage{
motif_enrichment(regions, annotations.list, file.label = NULL,
  pwm.bg = NULL, top.x = 20)
}
\arguments{
\item{regions}{The regions on which motif enrichment should be performed.}

\item{annotations.list}{A list of annotation databases returned by
\code{\link{select_annotations}}.}

\item{file.label}{A label for generating file names to save the results.
If \code{NULL}, results are not saved to disk.}

\item{pwm.bg}{A PWMLogn background object against which enrichment
should be performed.}

\item{top.x}{Number of top motifs for which logos are generated.}
}
\value{
A list with the following elements: \describe{
\item{Region}{The subset of regions where motifs were sought.}
\item{Enrichment}{The result of the \pkg{PWMEnrich} \code{\link[PWMEnrich]{motifEnrichment}} call.}
\item{Report}{The result of the \pkg{PWMEnrich} \code{\link[PWMEnrich]{groupReport}} call.}}
}
\description{
Performs motif enrichment against \code{pwm.bg}. If \code{pwm.bg} is not provided,
\code{annotations.list$PWMBG} is used.
}
