% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{motif_enrichment_genes}
\alias{motif_enrichment_genes}
\title{Perform motif enrichment on the promoters of a set of genes.}
\usage{
motif_enrichment_genes(selected.genes, annotations.list, flank.size = 1000,
  ...)
}
\arguments{
\item{selected.genes}{A vector of ENTREZ gene ids whose promoters should be
subjected to motif enrichment.}

\item{annotations.list}{A list of annotation databases returned by
\code{\link{select_annotations}}.}

\item{flank.size}{How many base pairs upstream of the TSS should we retrieve?}

\item{...}{Additional arguments for \code{\link{motif_enrichment}}.}
}
\value{
A list with the following elements: \describe{
  \item{Region}{The subset of regions where motifs were sought.}
  \item{Enrichment}{The result of the \pkg{PWMEnrich} \code{\link[PWMEnrich]{motifEnrichment call.}}}
  \item{Report}{The result of the \pkg{PWMEnrich} \code{\link[PWMEnrich]{groupReport call.}}}}
}
\description{
Utility function which performs the same operations as motif_enrichment,
but accepts a list of genes instead of a list of regions.
}

