% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsawAnalysis.R
\name{multiple.testing}
\alias{multiple.testing}
\title{Makes corrections for multiple testing}
\usage{
multiple.testing(filtered.counted.reads, results, txdb)
}
\arguments{
\item{filtered.counted.reads}{A filtered \linkS4class{RangedSummarizedExperiment} object returned by \code{\link[csaw]{WindowCounts}}.}

\item{results}{A \linkS4class{DGELRT} object returned as results by \code{\link[edgeR]{glmQLFTest}}.}
}
\value{
A list with the following elements: \describe{
  \item{[[1]]} {A \linkS4class{GRanges} object returned by \code{\link[csaw]{mergeWindows}}}
  \item{[[2]]} {A data frame returned by \code{\link[csaw]{combineTests}}}}
}
\description{
Makes corrections for multiple testing
}

