% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{multiple_region_enrichment}
\alias{multiple_region_enrichment}
\title{Performs region enrichment on a set of regions and returns summarized results.}
\usage{
multiple_region_enrichment(queries.regions, genome.regions,
  query.order = NULL, genome.order = NULL, file.prefix = NULL,
  plot.width = 7, plot.height = 7, individual.plots = FALSE)
}
\arguments{
\item{queries.regions}{A list of regions whose enrichment ratios must be calculated.}

\item{genome.regions}{The genome partition indicating which part of the genome
fall within which category. Each range should have a 'name' attribute}

\item{file.prefix}{An optional file name prefix for tables and graphical representation.}

\item{plot.width}{The width of any resulting summary plot.}

\item{plot.height}{The height of any resulting summary plot.}

\item{individual.plots}{If true, produce individual plots as well as combined plots.}

\item{factor.order}{An optional ordering of the region types for the produced plot.}
}
\value{
A list of summarized enrichment metrics.
}
\description{
Performs region enrichment on a set of regions and returns summarized results.
}
