% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsawAnalysis.R
\name{normalization.factors}
\alias{normalization.factors}
\title{Calculates normalization factors, by using TMM method.}
\usage{
normalization.factors(bam.files, param)
}
\arguments{
\item{bam.files}{A vector with the path to bam files.}

\item{param}{A vector with the list of parameters to give to the functions.}
}
\value{
A list with the following elements: \describe{
  \item{[[1]]} {A vector with the normalization factors to use to normalize.}
  \item{[[2]]} {A \linkS4class{RangedSummarizedExperiment} object returned by \code{\link[csaw]{WindowCounts}} to consider as a "bin".}}
}
\description{
Calculates normalization factors, by using TMM method.
}

