% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{number.of.components}
\alias{number.of.components}
\title{Return the number of components in a CHIA object.}
\usage{
number.of.components(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
The number of components in the chia object.
}
\description{
Return the number of components in a CHIA object.
}

