% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{number.of.contacts}
\alias{number.of.contacts}
\title{Return the number of contacts in a CHIA object.}
\usage{
number.of.contacts(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
The number of contacts in the chia object.
}
\description{
Return the number of contacts in a CHIA object.
}

