% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{pairwise.overlap}
\alias{pairwise.overlap}
\title{Calculate the pairwise overlaps of all factors within an intersect.object.}
\usage{
pairwise.overlap(intersect.object, filename = NULL)
}
\arguments{
\item{intersect.object}{An intersect object returned by build.intersect.}

\item{filename}{A name for prefix for the names of the output files where the
annotations will be written. Pass NULL to skip saving to a file.}
}
\value{
A matrix containing the pairwise overlaps of all factors in
}
\description{
Calculate the pairwise overlaps of all factors within an intersect.object.
}

