% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{partition_genomic_regions}
\alias{partition_genomic_regions}
\title{Generate a GRanges object partitioning the genome into Promoter, Exon, introns,
downstream and distal intergenic regions.}
\usage{
partition_genomic_regions(TxDb, available.genome = NULL, BSgenome = NULL,
  flank.width = 1000)
}
\arguments{
\item{TxDb}{A TxDb object giving the locations of all genes for the given genome.}

\item{available.genome}{A Granges object to be used as a background of all 
the available genome.}

\item{BSgenome}{A BSgenome object to be used for generating an available genome
background. Ignored if available.genome is provided.}

\item{flank.width}{The distance from the TSS/TES which are considered Promoters/Downstream.}
}
\value{
A GRanges object partitioning the genome into functional genomic regions.
}
\description{
Generate a GRanges object partitioning the genome into Promoter, Exon, introns,
downstream and distal intergenic regions.
}
