% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeChIA.R
\name{plot.network.heatmap}
\alias{plot.network.heatmap}
\title{Plot heatmaps in fonction of a ChIA-PET annotation}
\usage{
plot.network.heatmap(chia.obj, size.limit, variable.name, label, output.dir)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{size.limit}{The networks must have a size over size.limit.}

\item{variable.name}{Name of the column to use, or "TF" for transcription factors.}

\item{label}{The label to add to the heatmap title (name of the variable).}

\item{output.dir}{The name of the directory where output should be saved.}
}
\description{
Creates heatmaps of the proportion of a variable from the ChIA-PET annotation, given as parameter, in fonction of the networks.
To use the transcription factors as variable, write "TF". This will create the same heatmaps, but with presence or absence of each factor instead of proportions.
}

