% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsawAnalysis.R
\name{post_processing}
\alias{post_processing}
\title{Save results of \code{\link{multiple.testing}} with annotations}
\usage{
post_processing(merged, tabcom, txdb, orgdb)
}
\arguments{
\item{merged}{A \linkS4class{GRanges} object returned by \code{\link[csaw]{mergeWindows}}}

\item{tabcom}{A data frame returned by \code{\link[csaw]{combineTests}}}
}
\value{
annotations, list, annotation of considered regions
}
\description{
Save results of \code{\link{multiple.testing}} with annotations
}

