% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{project.ranges}
\alias{project.ranges}
\title{Projects the ranges in query into the ranges in target.}
\usage{
project.ranges(query, target)
}
\arguments{
\item{query}{The ranges to be projected.}

\item{target}{The ranges to be projected against.}
}
\value{
The projection of the query ranges on the target ranges.
}
\description{
Returns the ranges in target overlapping the ranges in query, adjusting
their boundaries so that only the overlapping parts of the target ranges 
are returned.
}

