% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{proportion.genes}
\alias{proportion.genes}
\title{Return the proportion of nodes representing genes in a CHIA object.}
\usage{
proportion.genes(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
The number of active genes represented in the chia object.
}
\description{
Return the proportion of nodes representing genes in a CHIA object.
}

