% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsawAnalysis.R
\name{reads.to.counts}
\alias{reads.to.counts}
\title{Convert reads from bam files to counts}
\usage{
reads.to.counts(bam.files, param)
}
\arguments{
\item{bam.files}{A vector with the path to bam files.}

\item{param}{A vector with the list of parameters to give to the functions.}
}
\value{
A list with the following elements: \describe{
  \item{[[1]]} {A \linkS4class{RangedSummarizedExperiment} object returned by \code{\link[csaw]{WindowCounts}}.}
  \item{[[2]]} {A data frame with the necessary information for the construction of a plot of window sizes.}
  \item{[[3]]} {A vector with correlated reads.}}
}
\description{
Convert reads from bam files to counts
}

