% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{region.enrichment.summary}
\alias{region.enrichment.summary}
\title{Performs a summary of region enrichment results.}
\usage{
region.enrichment.summary(result.list, file.prefix = NULL, plot.width = 7,
  plot.height = 7)
}
\arguments{
\item{result.list}{a list of results returned by region.enrichment.}

\item{file.prefix}{An optional file name prefix for tables and graphical representation.}

\item{plot.width}{The width of any resulting plot.}

\item{plot.height}{The height of any resulting plot.}
}
\value{
A list of summarized enrichment metrics.
}
\description{
Performs a summary of region enrichment results.
}

