% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{region.overlap}
\alias{region.overlap}
\title{Generate a matrix detailing the percentage of overlap between query regions and
a list of target regions.}
\usage{
region.overlap(query.regions, target.regions, proportion = TRUE)
}
\arguments{
\item{query.regions}{The GRanges object detailing the regions against which 
the overlaps must be calculated.}

\item{target.regions}{A GRangesList object detailing the regions whose overlaps
against each region in query.regions must be calculated.}

\item{proportion}{If TRUE, proportions are returned instead of base pair counts.}
}
\value{
A matrix detailing the overlap between teh query and the targets.
}
\description{
Generate a matrix detailing the percentage of overlap between query regions and
a list of target regions.
}

