% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{region_consensus}
\alias{region_consensus}
\title{Obtains consensus regions from a set of regions.}
\usage{
region_consensus(regions, keep.signal = TRUE, fake.signal = FALSE)
}
\arguments{
\item{regions}{The regiosn to be consensus'ed.}

\item{keep.signal}{Whether or not the signal should be kept.}
}
\value{
The collapsed regions.
}
\description{
Obtains consensus regions from a set of regions.
}

