% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{region_enrichment}
\alias{region_enrichment}
\title{Calculates enrichment ratios for quer regions against a genome wide
partition of the genome.}
\usage{
region_enrichment(query.regions, genome.wide, genome.order = NULL,
  file.out = NULL)
}
\arguments{
\item{query.regions}{The regions whose enrichment ratios must be calculated.}

\item{genome.wide}{The genome partition indicating which part of the genome
fall within which category. Each range should have a 'name' attribute
indicating its category.}

\item{file.out}{An optional file name for a graphical representation of the enrichments.}

\item{factor.order}{An optional ordering of the region types for the produced plot.}
}
\value{
A data-frame containing the enrichment values.
}
\description{
Calculates enrichment ratios for quer regions against a genome wide
partition of the genome.
}

