% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{region_enrichment_summary}
\alias{region_enrichment_summary}
\title{Performs a summary of region enrichment results.}
\usage{
region_enrichment_summary(result.list, file.prefix = NULL,
  query.order = NULL, genome.order = NULL, plot.width = 7,
  plot.height = 7)
}
\arguments{
\item{result.list}{a list of results returned by region_enrichment.}

\item{file.prefix}{An optional file name prefix for tables and graphical representation.}

\item{plot.width}{The width of any resulting plot.}

\item{plot.height}{The height of any resulting plot.}

\item{genome.regions}{The genome partition indicating which part of the genome
fall within which category. Each range should have a 'name' attribute}

\item{factor.order}{An optional ordering of the region types for the produced plot.}
}
\value{
A list of summarized enrichment metrics.
}
\description{
Performs a summary of region enrichment results.
}

