% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{select.annotations}
\alias{select.annotations}
\title{Annotations selection helper.}
\usage{
select.annotations(genome.build)
}
\arguments{
\item{genome.build}{The genome build to use for annotation.}
}
\value{
A list with the following elements:
  - *TxDb*: A TxDb of class AnnotationDBI providing informations about the genes
    of the selected build.
  - *OrgDb*: An OrgDb object for the selected species.
  - *OrgDbStr*: A character string representing the name of the OrgDb object.
  - *BSGenome*: A BSGenome object to retrieve DNA sequences
  - *KEGG*: A cache of KEGG pathways, as returned by kegg.gsets from the gage library.
  - *PWMBG*: A PWMLogn background for motif enrichment of promoter regions.
}
\description{
Given a genome build identifier, creates a list of annotation databases
which can be passed to the annotation functions.
}

