% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{separate.into.communities}
\alias{separate.into.communities}
\title{Separetes a large networks into communities and saves the sub-netwoks in cytoscape-friendly format}
\usage{
separate.into.communities(network.input, network, annotated.chia, output.dir,
  method = igraph::cluster_fast_greedy)
}
\arguments{
\item{network.input}{The csv file containing the network to divide.}

\item{network}{The id of the network.}

\item{annotated.chia}{The txt file containing all information about the networks.}

\item{output.dir}{The directory where to save the files.}

\item{method}{The algorithm to use to divide the graph.}
}
\description{
Separetes a large networks into communities and saves the sub-netwoks in cytoscape-friendly format
}

