% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{simple.chrom.state.map}
\alias{simple.chrom.state.map}
\title{Returns the mapping used to go from full chromatin states to simple ones.}
\usage{
simple.chrom.state.map()
}
\value{
The mapping used to go from full chromatin states to simple ones.
}
\description{
Returns the mapping used to go from full chromatin states to simple ones.
}

