% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{simple_chrom_state_map}
\alias{simple_chrom_state_map}
\title{Returns the mapping used to go from full chromatin states to simple ones.}
\usage{
simple_chrom_state_map()
}
\value{
The mapping used to go from full chromatin states to simple ones.
}
\description{
Returns the mapping used to go from full chromatin states to simple ones.
}
