% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{simplify.chrom.state}
\alias{simplify.chrom.state}
\title{Simplifies chromatin state names.}
\usage{
simplify.chrom.state(chrom.state.names)
}
\arguments{
\item{chrom.state.names}{The chromatin states to be simplified.}
}
\value{
The simplified chromatin states.
}
\description{
Given a vector of canonical 18-state HMM-derived chromatin states,
return a vector of the same length where each state is associated
with a simplified version of itself.
}

