% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{split.by.community}
\alias{split.by.community}
\title{Associates components ids and sizes to chia data, as returned by \code{\link{load.chia}}.}
\usage{
split.by.community(chia.obj, oneByOne = FALSE,
  method = igraph::cluster_fast_greedy, weight.attr = NULL)
}
\arguments{
\item{chia.obj}{ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{oneByOne}{Sould the netwoks be treated one by one or as a whole?}

\item{method}{What method sould be used to split data (ignored if split = \code{FALSE}).}

\item{split}{Should the data be divided into communities?}
}
\value{
The annotated chia.obj.
}
\description{
Associates components ids and sizes to chia data, as returned by \code{\link{load.chia}}.
}

