% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsawAnalysis.R
\name{test.diff.binding}
\alias{test.diff.binding}
\title{Tests the differential binding.}
\usage{
test.diff.binding(bam.files, filtered.counted.reads, norm.factors,
  correspondances, reference)
}
\arguments{
\item{bam.files}{A vector with the path to bam files.}

\item{filtered.counted.reads}{A filtered \linkS4class{RangedSummarizedExperiment} object returned by \code{\link[csaw]{WindowCounts}}.}

\item{norm.factors}{A vector with the normalization factors to use to normalize.}
}
\value{
A \linkS4class{DGELRT} object returned as results by \code{\link[edgeR]{glmQLFTest}}.
}
\description{
Tests the differential binding.
}

