% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{analyze.chia.pet}
\alias{analyze.chia.pet}
\title{Analyze ChIA-PET data and produce graphs.}
\usage{
analyze.chia.pet(input.chia, input.chrom.state = NULL, biosample = NULL,
  genome.build = NULL, tf.regions = NULL, histone.regions = NULL,
  pol.regions = NULL, expression.data = NULL, output.dir = "output",
  verbose = TRUE, ...)
}
\arguments{
\item{input.chia}{The file containing processed ChIA-PET data.}

\item{input.chrom.state}{The name of the file containing the information about chromatin states.}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are GM12878, K562 or MCF-7.}

\item{genome.build}{The name of the chosen annotation ("hg38", "hg19").}

\item{tf.regions}{A \linkS4class{GRangesList} object containing TF regions.}

\item{histone.regions}{A \linkS4class{GRangesList} object containing histone regions.}

\item{pol.regions}{A \linkS4class{GRangesList} object containing the pol2 regions.}

\item{output.dir}{The name of the directory where output should be saved.}

\item{...}{Additional parameters to be passed to \code{\link{annotate.chia}}}

\item{expression.levels}{A data frame containing the levels of expression of genes, according to their EMSEMBL id.}
}
\value{
The annotated chia.obj.
}
\description{
Analyze ChIA-PET data and produce graphs.
}

