% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeChIA.R
\name{analyze.chia.pet}
\alias{analyze.chia.pet}
\title{Performs all possible analysis steps on a ChIA-PET object.}
\usage{
analyze.chia.pet(chia.obj, output.dir = ".", verbose = TRUE)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{output.dir}{The directory where output should be saved.}

\item{verbose}{Set to TRUE to get progress output to the console.}

\item{label}{The label to add to the heatmap title (name of the variable).}
}
\description{
Performs all possible analysis steps on a ChIA-PET object.
}

