% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{analyze.chia.pet}
\alias{analyze.chia.pet}
\title{Analyze ChIA-PET data and produce graphs.}
\usage{
<<<<<<< HEAD
analyze.chia.pet(input.chia, input.chrom.state = NULL, biosample = NULL,
  genome.build = NULL, tf.regions = NULL, histone.regions = NULL,
  expression.data = NULL, output.dir = "output/")
}
\arguments{
\item{input.chia}{The file containing processed ChIA-PET data.}

\item{input.chrom.state}{The name of the file containing the information about chromatin states.}
=======
analyze.chia.pet(input.chia, biosample = NULL, genome.build = NULL,
  output.dir = "output/", reset.cache = TRUE)
}
\arguments{
\item{input.chia}{The path of the file containing the ChIA-PET data.}
>>>>>>> 0eca4f25f233ad486ade10b9943ede4f990067af

\item{biosample}{The biosample identifier from ENCODE. Valid examples are GM12878, K562 or MCF-7.}

\item{genome.build}{The name of the chosen annotation ("hg38", "hg19").}

\item{tf.regions}{A data frame containing the TF data.}

\item{histone.regions}{A data frame containing the histone data.}

\item{output.dir}{The name of the directory where to save the graphs.}

<<<<<<< HEAD
\item{expression.levels}{A data frame containing the levels of expression of genes. according to their EMSEMBL id.}
}
\value{
The annotated chia.obj.
=======
\item{input.chrom.state}{The name of the file containing the information about chromatin states.}
>>>>>>> 0eca4f25f233ad486ade10b9943ede4f990067af
}
\description{
Analyze ChIA-PET data and produce graphs.
}

