% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{analyze.chia.pet}
\alias{analyze.chia.pet}
\title{Analyze ChIA-PET data and produce graphs.}
\usage{
analyze.chia.pet(input.chia, biosample = NULL, genome.build = NULL,
  output.dir = "output/", reset.cache = TRUE)
}
\arguments{
\item{input.chia}{The path of the file containing the ChIA-PET data.}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are GM12878, K562 or MCF-7.}

\item{genome.build}{The name of the chosen annotation ("hg38", "hg19").}

\item{output.dir}{The name of the directory where to save the graphs.}

\item{input.chrom.state}{The name of the file containing the information about chromatin states.}
}
\description{
Analyze ChIA-PET data and produce graphs.
}

