% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{annotate.chia}
\alias{annotate.chia}
\title{Annotate "chia.obj", given as parameter.}
\usage{
annotate.chia(chia.obj, input.chrom.state, tf.regions, histone.regions,
  expression.levels, genome.build = c("hg19", "mm9", "mm10", "hg38"),
  biosample = "GM12878", tssRegion, output.dir)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{input.chrom.state}{The name of the file containing the information about chromatin states.}

\item{tf.regions}{A data frame containing the TF data.}

\item{histone.regions}{A data frame containing the histone data.}

\item{expression.levels}{A data frame containing the levels of expression of genes. according to their EMSEMBL id.}

\item{genome.build}{The name of the chosen annotation ("hg38", "mm9", "mm10", "hg19").}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562.}

\item{tssRegion}{A vector with the region range to TSS.}

\item{output.dir}{The name of the directory where to write the selected annotations.}
}
\value{
The annotated "\code{chia.obj}".
}
\description{
Annotate "chia.obj", given as parameter.
}

