% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{annotate.chia}
\alias{annotate.chia}
\title{Annotate "chia.obj", given as parameter.}
\usage{
annotate.chia(chia.obj, input.chrom.state, tf.regions, histone.regions,
  expression.levels, genome.build = c("hg19", "mm9", "mm10", "hg38"),
<<<<<<< HEAD
  biosample = "GM12878", output.dir)
=======
  output.dir)
>>>>>>> 0eca4f25f233ad486ade10b9943ede4f990067af
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{input.chrom.state}{The name of the file containing the information about chromatin states.}

\item{tf.regions}{A data frame containing the TF data.}

<<<<<<< HEAD
\item{histone.regions}{A data frame containing the histone data.}
=======
\item{histone.regions}{A \linkS4class{GRangesList} object defining histone regions.}
>>>>>>> 0eca4f25f233ad486ade10b9943ede4f990067af

\item{expression.levels}{A data frame containing the levels of expression of genes. according to their EMSEMBL id.}

\item{genome.build}{The name of the chosen annotation ("hg38", "mm9", "mm10", "hg19").}

<<<<<<< HEAD
\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562.}

=======
>>>>>>> 0eca4f25f233ad486ade10b9943ede4f990067af
\item{output.dir}{The name of the directory where to write the selected annotations.}
}
\value{
The annotated "\code{chia.obj}".
}
\description{
Annotate "chia.obj", given as parameter.
}

