% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{annotate.venn.center}
\alias{annotate.venn.center}
\title{Annotate the inner group of an intersect.object.}
\usage{
annotate.venn.center(intersect.object, filename = NULL)
}
\arguments{
\item{intersect.object}{An intersect object returned by build.intersect.}

\item{filename}{A filename for the resulting annotations. Pass NULL
to skip saving to a file.}
}
\value{
The annotations for the intersect's inner regions.
}
\description{
Given an intersect.object, generate annotations for the regions
where all factors are present.
}

