% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntersectionHelper.R
\name{annotate.venn.exclusive}
\alias{annotate.venn.exclusive}
\title{Annotate the outer groups of an intersect.object.}
\usage{
annotate.venn.exclusive(intersect.object, file.prefix = NULL)
}
\arguments{
\item{intersect.object}{An intersect object returned by build.intersect.}

\item{file.prefix}{A prefix for the names of the output files where the
annotations will be written. Pass NULL to skip saving to a file.}
}
\value{
A list with the generated annotations.
}
\description{
Given an intersect.object, generate annotations for the regions
where only one factor is present. One annotation per factor is generated.
}

