% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{associate.chrom.state}
\alias{associate.chrom.state}
\title{Associate chromatin sates with the \code{Regions} of "chia.obj".}
\usage{
associate.chrom.state(chia.obj, input.chrom.state)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{input.chrom.state}{The path to a bed file representing a genome 
segmentation by chromatin state.}
}
\value{
"\code{chia.obj}" with associated chromatin states.
}
\description{
Associate chromatin sates with the \code{Regions} of "chia.obj".
}

