% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R, R/CharacterizeChIA.R
\name{associate.chrom.state}
\alias{associate.chrom.state}
\title{Associate chromatin sates to a \linkS4class{GRanges} object.}
\usage{
associate.chrom.state(chia.obj, input.chrom.state)

associate.chrom.state(chia.obj, input.chrom.state)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{input.chrom.state}{The path to a bed file representing a genome
segmentation by chromatin state.}

\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{input.chrom.state}{The path to a bed file representing a genome
segmentation by chromatin state.}
}
\value{
The \linkS4class{GRanges} object with associated chromatin states.

"\code{chia.obj}" with associated chromatin states.
}
\description{
Associate chromatin sates to a \linkS4class{GRanges} object.

Associate chromatin sates with the \code{Regions} of "chia.obj".
}

