% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{associate.chrom.state}
\alias{associate.chrom.state}
\title{Associate chromatin sates with the \code{Regions} of "chia.obj".}
\usage{
associate.chrom.state(chia.obj, input.chrom.states)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{input.chrom.sates}{The name of the file containing the information about chromatin states.}
}
\value{
"\code{chia.obj}" with associated chromatin states.
}
\description{
Associate chromatin sates with the \code{Regions} of "chia.obj".
}

