% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{associate.fitness.genes}
\alias{associate.fitness.genes}
\title{Identify essential genes of the \code{Regions} of "chia.obj".}
\usage{
associate.fitness.genes(chia.obj, cell.type)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{cell.type}{The type of the considered cell.}
}
\value{
"\code{chia.obj}" with identified essential genes.
}
\description{
Identify essential genes of the \code{Regions} of "chia.obj".
}

