% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate.fitness.genes}
\alias{associate.fitness.genes}
\title{Identify essential genes of a \linkS4class{GRanges} object.}
\usage{
associate.fitness.genes(regions)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}
}
\value{
The \linkS4class{GRanges} object with identified essential genes.
}
\description{
Identify essential genes of a \linkS4class{GRanges} object.
}

