% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{associate.gene}
\alias{associate.gene}
\title{Associate genes with the \code{Regions} of "chia.obj".}
\usage{
associate.gene(chia.obj, expression.data = NULL)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{expression.data}{A data frame containing the levels of expression of genes, according to their EMSEMBL id.}
}
\value{
"\code{chia.obj}" with associated genes.
}
\description{
Associate genes with the \code{Regions} of "chia.obj".
}

