% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R, R/CharacterizeChIA.R
\name{associate.gene}
\alias{associate.gene}
\title{Associate genes to a \linkS4class{GRanges} object from ChIA-PET data.}
\usage{
associate.gene(chia.obj, expression.data = NULL)

associate.gene(chia.obj, expression.data = NULL)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{expression.data}{A data frame containing the levels of expression of genes, according to their EMSEMBL id.}

\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{expression.data}{A data frame containing the levels of expression of genes, according to their EMSEMBL id.}
}
\value{
The \linkS4class{GRanges} object with associated genes.

"\code{chia.obj}" with associated genes.
}
\description{
Associate genes to a \linkS4class{GRanges} object from ChIA-PET data.

Associate genes with the \code{Regions} of "chia.obj".
}

