% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{associate.histone.marks}
\alias{associate.histone.marks}
\title{Associate histone marks with the \code{Regions} of "chia.obj".}
\usage{
associate.histone.marks(chia.obj, genome.build, biosample)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by load.chia.}

\item{genome.build}{The name of the chosen annotation ("hg38", "mm9", "mm10", "hg19").}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or MCF-7.}
}
\value{
"\code{chia.obj}" with associated histone overlap percentage.
}
\description{
Associate histone marks with the \code{Regions} of "chia.obj".
}

