% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R
\name{associate.histone.marks}
\alias{associate.histone.marks}
\title{Associate histone marks to a \linkS4class{GRanges} object.}
\usage{
associate.histone.marks(regions, histone.regions)
}
\arguments{
\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{genome.build}{The name of the chosen annotation ("hg38", "mm9", "mm10", "hg19").}

\item{biosample}{The biosample identifier from ENCODE. Valid examples are
GM12878, K562 or MCF-7.}
}
\value{
The \linkS4class{GRanges} object with associated histone overlap percentage.
}
\description{
Associate histone marks to a \linkS4class{GRanges} object.
}

