% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHelper.R, R/CharacterizeChIA.R
\name{associate.tf}
\alias{associate.tf}
\title{Associate TF overlaps to a \linkS4class{GRanges} object.}
\usage{
associate.tf(chia.obj, tf.regions)

associate.tf(chia.obj, tf.regions)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{tf.regions}{A data frame containing the TF data.}

\item{regions}{A \linkS4class{GRanges} object to annotate.}

\item{tf.regions}{A data frame containing the TF data.}
}
\value{
The \linkS4class{GRanges} object with associated TF overlaps.

"\code{chia.obj}" with associated TF overlaps.
}
\description{
Associate TF overlaps to a \linkS4class{GRanges} object.

Associate TF overlaps with the \code{Regions} of "chia.obj".
}

