% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeHelper.R
\name{default.download.filter.rna}
\alias{default.download.filter.rna}
\title{Default filtering function for download.encode.rna.}
\usage{
default.download.filter.rna(query.results, genome.assembly)
}
\arguments{
\item{query.results}{A partial data.frame obtained from the queryEncode
function.The biosample identifier from ENCODE. Valid examples are}

\item{genome.assembly}{Which genome assembly should the results come from?}
}
\value{
A filtered data frame.
}
\description{
The filtering function does three things:
  - It removes all files which do not have the correct genome assembly.
  - It only keeps files with counts for genes, not transcripts or isoforms.
  - If removes any file where treatment is not NA.
}

