% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeChIA.R
\name{output.annotated.chia}
\alias{output.annotated.chia}
\title{Save annotated ChIA-PET data.}
\usage{
output.annotated.chia(chia.obj, chia.raw, output.dir = "output")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{chia.raw}{The raw ChIA-PET data, before annotation.}

\item{output.dir}{The name of the directory where to save the files.}
}
\description{
Writes files: \itemize{
  \item The first one separates the "left" and "right" sides of the ChIA-PET data with annotations.
  \item The second set of files is made of components files. They group all interactions in a single component, with
       an extra column containing the number of reads supporting the data. Their format is supported by Cytoscape.
  \item The final file contains the annotated Regions of the ChIA-PET data, with two extra columns: \describe{
       \item{$Component.Id} {Id (number) of the hub in which the node is found.}
       \item{$Componend.Size} {Number of nodes of the component in which the node is found.}}}
}

